SOFB.MNT: create module for BPM class.
